* Fix KAI compilation. * Fix 3 sec countdown before game starts, which was broken when server ran minimised.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2905 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/GameSetup.h
@@ -27,7 +27,7 @@ public:
 	int hostport;
 	int sourceport;			//the port clients will try to connect from
 
-	float readyTime;
+	int readyTime;
 	bool forceReady;
 
 	bool Init(char* buf, int size);
